adding group comm folder and organizing mocks in there
